Add Product Details Screen and Update Router
Added a new Product Details screen along with its associated widgets (ColorDot and SizeButton). Updated the AppRouter to include the new route for the Product Details screen and modified the FilterResultGridView to navigate to this screen when a product is tapped. Also, updated the SectionHeader widget to accept a dynamic text parameter.